+++ sketch-core/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/core/android/test/decode/internal/DrawableDecoderTest.kt
@@ -17,30 +17,42 @@
 package com.github.panpf.sketch.core.android.test.decode.internal
 
 import android.content.res.Resources
+import android.graphics.ColorSpace
 import android.os.Build
+import android.os.Build.VERSION
+import android.os.Build.VERSION_CODES
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.panpf.sketch.ColorType
+import com.github.panpf.sketch.colorType
 import com.github.panpf.sketch.decode.internal.DrawableDecoder
+import com.github.panpf.sketch.decode.internal.createResizeTransformed
 import com.github.panpf.sketch.decode.internal.createScaledTransformed
 import com.github.panpf.sketch.drawable.ResDrawable
 import com.github.panpf.sketch.fetch.FetchResult
-import com.github.panpf.sketch.fetch.copy
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.request.colorSpace
+import com.github.panpf.sketch.request.colorType
+import com.github.panpf.sketch.resize.Precision
+import com.github.panpf.sketch.resize.Resize
+import com.github.panpf.sketch.resize.Scale
+import com.github.panpf.sketch.size
 import com.github.panpf.sketch.source.DataFrom.LOCAL
 import com.github.panpf.sketch.source.DrawableDataSource
 import com.github.panpf.sketch.test.singleton.getTestContextAndSketch
+import com.github.panpf.sketch.test.utils.TestColor
+import com.github.panpf.sketch.test.utils.colorSpaceNameCompat
+import com.github.panpf.sketch.test.utils.corners
+import com.github.panpf.sketch.test.utils.decode
 import com.github.panpf.sketch.test.utils.getBitmapOrThrow
-import com.github.panpf.sketch.test.utils.shortInfoColorSpace
 import com.github.panpf.sketch.test.utils.toDecoder
 import com.github.panpf.sketch.test.utils.toRequestContext
 import com.github.panpf.sketch.util.Size
-import com.github.panpf.sketch.util.toShortInfoString
+import com.github.panpf.sketch.util.computeScaleMultiplierWithOneSide
 import com.github.panpf.tools4a.dimen.ktx.dp2px
 import kotlinx.coroutines.test.runTest
 import org.junit.runner.RunWith
-import kotlin.math.min
-import kotlin.math.roundToInt
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -109,29 +121,26 @@ class DrawableDecoderTest {
         val (context, sketch) = getTestContextAndSketch()
         val factory = DrawableDecoder.Factory()
 
-        val resourceUri = newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.test)
-        ImageRequest(context, resourceUri)
-            .toDecoder(sketch, factory)
+        val imageUri = newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.test)
+        val imageSize = Size(60.dp2px, 30.dp2px)
+        val request = ImageRequest(context, imageUri)
+
+        request.toDecoder(sketch, factory)
             .imageInfo.apply {
-                val imageWidth = 50.dp2px
-                val imageHeight = 40.dp2px
                 assertEquals(
-                    "ImageInfo(${imageWidth}x${imageHeight},'text/xml')",
-                    toShortString()
+                    expected = "ImageInfo($imageSize,'text/xml')",
+                    actual = toShortString()
                 )
             }
 
-        ImageRequest(context, resourceUri)
-            .toDecoder(sketch, factory) {
-                it.copy(mimeType = null)
-            }.imageInfo.apply {
-                val imageWidth = 50.dp2px
-                val imageHeight = 40.dp2px
-                assertEquals(
-                    "ImageInfo(${imageWidth}x${imageHeight},'image/png')",
-                    toShortString()
-                )
-            }
+        request.toDecoder(sketch, factory) { fetchResult ->
+            fetchResult.copy(mimeType = null)
+        }.imageInfo.apply {
+            assertEquals(
+                expected = "ImageInfo($imageSize,'image/png')",
+                actual = toShortString()
+            )
+        }
 
         ImageRequest(context, newResourceUri(8801)).run {
             assertFailsWith(Resources.NotFoundException::class) {
@@ -155,72 +164,163 @@ class DrawableDecoderTest {
         val (context, sketch) = getTestContextAndSketch()
 
         val factory = DrawableDecoder.Factory()
-        val imageWidth = 50.dp2px
-        val imageHeight = 40.dp2px
+        val imageUri = newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.test)
+        val imageSize = Size(60.dp2px, 30.dp2px)
+        val request = ImageRequest(context, imageUri)
 
-        ImageRequest(
-            context,
-            newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.test)
-        ) {
-            resize(imageWidth / 2, imageWidth / 2)
-        }.run {
-            val fetcher = sketch.components.newFetcherOrThrow(
-                this
-                    .toRequestContext(sketch, Size.Empty)
-            )
-            val fetchResult = fetcher.fetch().getOrThrow()
-            factory.create(this@run.toRequestContext(sketch), fetchResult)!!.decode()
-        }.apply {
-            val scale = min(
-                (imageWidth / 2) / imageWidth.toFloat(),
-                (imageWidth / 2) / imageHeight.toFloat()
-            )
-            val scaledSize = Size(
-                width = (imageWidth * scale).roundToInt(),
-                height = (imageHeight * scale).roundToInt()
-            )
+        // default
+        request.newRequest {
+            size(Size.Origin)
+        }.decode(sketch, factory).apply {
+            val bitmap = image.getBitmapOrThrow()
+            assertEquals(expected = imageSize, actual = bitmap.size)
+            assertEquals(expected = ColorType.ARGB_8888, actual = bitmap.colorType)
+            assertEquals(expected = colorSpaceNameCompat(), actual = bitmap.colorSpaceNameCompat)
+            assertEquals(expected = listOf(0, 0, 0, 0), actual = bitmap.corners())
+            assertEquals(expected = null, actual = transformeds)
             assertEquals(
-                "Bitmap(${scaledSize},ARGB_8888${shortInfoColorSpace("SRGB")})",
-                image.getBitmapOrThrow().toShortInfoString()
-            )
-            assertEquals(listOf(createScaledTransformed(scale)), transformeds)
-            assertEquals(
-                "ImageInfo(${imageWidth}x${imageHeight},'text/xml')",
-                imageInfo.toShortString()
+                expected = "ImageInfo($imageSize,'text/xml')",
+                actual = imageInfo.toShortString()
             )
             assertEquals(LOCAL, dataFrom)
-        }
+        }.image.getBitmapOrThrow()
 
-        ImageRequest(
-            context,
-            newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.test)
-        ) {
-            resize(imageWidth * 2, imageWidth * 2)
-        }.run {
-            val fetcher = sketch.components.newFetcherOrThrow(
-                this.toRequestContext(sketch, Size.Empty)
-            )
-            val fetchResult = fetcher.fetch().getOrThrow()
-            factory.create(this@run.toRequestContext(sketch), fetchResult)!!.decode()
-        }.apply {
-            val scale = min(
-                (imageWidth * 2) / imageWidth.toFloat(),
-                (imageWidth * 2) / imageHeight.toFloat()
-            )
-            val scaledSize = Size(
-                width = (imageWidth * scale).roundToInt(),
-                height = (imageHeight * scale).roundToInt()
+        // resize: scale, EXACTLY, START_CROP
+        val targetSize = Size(100, 100)
+        val targetScale = computeScaleMultiplierWithOneSide(
+            sourceSize = imageSize,
+            targetSize = targetSize
+        )
+        var targetResize =
+            Resize(size = targetSize, precision = Precision.EXACTLY, scale = Scale.START_CROP)
+        request.newRequest {
+            resize(targetResize.size, targetResize.precision, targetResize.scale)
+        }.decode(sketch, factory).apply {
+            val bitmap = image.getBitmapOrThrow()
+            assertEquals(expected = targetSize, actual = bitmap.size)
+            assertEquals(expected = ColorType.ARGB_8888, actual = bitmap.colorType)
+            assertEquals(expected = colorSpaceNameCompat(), actual = bitmap.colorSpaceNameCompat)
+            assertEquals(
+                expected = listOf(0, 1342177280, 1342177280, 0),
+                actual = bitmap.corners()
             )
             assertEquals(
-                "Bitmap(${scaledSize},ARGB_8888${shortInfoColorSpace("SRGB")})",
-                image.getBitmapOrThrow().toShortInfoString()
+                expected = listOf(
+                    createScaledTransformed(targetScale),
+                    createResizeTransformed(targetResize)
+                ),
+                actual = transformeds
             )
-            assertEquals(listOf(createScaledTransformed(2.0f)), transformeds)
             assertEquals(
-                "ImageInfo(${imageWidth}x${imageHeight},'text/xml')",
-                imageInfo.toShortString()
+                expected = "ImageInfo($imageSize,'text/xml')",
+                actual = imageInfo.toShortString()
             )
             assertEquals(LOCAL, dataFrom)
+        }.image.getBitmapOrThrow()
+
+        // resize: scale, EXACTLY, CENTER_CROP
+        targetResize =
+            Resize(size = targetSize, precision = Precision.EXACTLY, scale = Scale.CENTER_CROP)
+        request.newRequest {
+            resize(targetResize.size, targetResize.precision, targetResize.scale)
+        }.decode(sketch, factory).apply {
+            val bitmap = image.getBitmapOrThrow()
+            assertEquals(expected = targetSize, actual = bitmap.size)
+            assertEquals(expected = ColorType.ARGB_8888, actual = bitmap.colorType)
+            assertEquals(expected = colorSpaceNameCompat(), actual = bitmap.colorSpaceNameCompat)
+            assertEquals(
+                expected = listOf(1342177280, 1342177280, 1342177280, 1342177280),
+                actual = bitmap.corners()
+            )
+            assertEquals(
+                expected = listOf(
+                    createScaledTransformed(targetScale),
+                    createResizeTransformed(targetResize)
+                ),
+                actual = transformeds
+            )
+            assertEquals(
+                expected = "ImageInfo($imageSize,'text/xml')",
+                actual = imageInfo.toShortString()
+            )
+            assertEquals(LOCAL, dataFrom)
+        }.image.getBitmapOrThrow()
+
+        // size: scale, EXACTLY, END_CROP
+        targetResize =
+            Resize(size = targetSize, precision = Precision.EXACTLY, scale = Scale.END_CROP)
+        request.newRequest {
+            resize(targetResize.size, targetResize.precision, targetResize.scale)
+        }.decode(sketch, factory).apply {
+            val bitmap = image.getBitmapOrThrow()
+            assertEquals(expected = targetSize, actual = bitmap.size)
+            assertEquals(expected = ColorType.ARGB_8888, actual = bitmap.colorType)
+            assertEquals(expected = colorSpaceNameCompat(), actual = bitmap.colorSpaceNameCompat)
+            assertEquals(
+                expected = listOf(1342177280, 0, 0, 1342177280),
+                actual = bitmap.corners()
+            )
+            assertEquals(
+                expected = listOf(
+                    createScaledTransformed(targetScale),
+                    createResizeTransformed(targetResize)
+                ),
+                actual = transformeds
+            )
+            assertEquals(
+                expected = "ImageInfo($imageSize,'text/xml')",
+                actual = imageInfo.toShortString()
+            )
+            assertEquals(LOCAL, dataFrom)
+        }.image.getBitmapOrThrow()
+
+        // colorType
+        request.newRequest {
+            size(Size.Origin)
+            colorType(ColorType.RGB_565)
+        }.decode(sketch, factory).apply {
+            val bitmap = image.getBitmapOrThrow()
+            assertEquals(expected = imageSize, actual = bitmap.size)
+            assertEquals(expected = ColorType.RGB_565, actual = bitmap.colorType)
+            assertEquals(expected = colorSpaceNameCompat(), actual = bitmap.colorSpaceNameCompat)
+            assertEquals(
+                expected = listOf(
+                    TestColor.BLACK,
+                    TestColor.BLACK,
+                    TestColor.BLACK,
+                    TestColor.BLACK
+                ),
+                actual = bitmap.corners()
+            )
+            assertEquals(expected = null, actual = transformeds)
+            assertEquals(
+                expected = "ImageInfo($imageSize,'text/xml')",
+                actual = imageInfo.toShortString()
+            )
+            assertEquals(LOCAL, dataFrom)
+        }.image.getBitmapOrThrow()
+
+        // colorSpace
+        if (VERSION.SDK_INT >= VERSION_CODES.O) {
+            request.newRequest {
+                size(Size.Origin)
+                colorSpace(ColorSpace.Named.DISPLAY_P3)
+            }.decode(sketch, factory).apply {
+                val bitmap = image.getBitmapOrThrow()
+                assertEquals(expected = imageSize, actual = bitmap.size)
+                assertEquals(expected = ColorType.ARGB_8888, actual = bitmap.colorType)
+                assertEquals(
+                    expected = colorSpaceNameCompat("DISPLAY_P3"),
+                    actual = bitmap.colorSpaceNameCompat
+                )
+                assertEquals(expected = listOf(0, 0, 0, 0), actual = bitmap.corners())
+                assertEquals(expected = null, actual = transformeds)
+                assertEquals(
+                    expected = "ImageInfo($imageSize,'text/xml')",
+                    actual = imageInfo.toShortString()
+                )
+                assertEquals(LOCAL, dataFrom)
+            }.image.getBitmapOrThrow()
         }
 
         ImageRequest(context, newResourceUri(8801)).run {
@@ -238,5 +338,7 @@ class DrawableDecoderTest {
                 )!!.decode()
             }
         }
+
+        // TODO test colorType and colorSpace
     }
 }