+++ sketch-core/src/androidMain/kotlin/com/github/panpf/sketch/util/drawables.kt
@@ -20,8 +20,8 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.content.res.Resources
 import android.graphics.Bitmap
-import android.graphics.Bitmap.Config
 import android.graphics.Canvas
+import android.graphics.ColorSpace
 import android.graphics.Rect
 import android.graphics.drawable.AnimatedImageDrawable
 import android.graphics.drawable.AnimatedVectorDrawable
@@ -33,7 +33,7 @@ import android.graphics.drawable.DrawableWrapper
 import android.graphics.drawable.GradientDrawable
 import android.graphics.drawable.TransitionDrawable
 import android.graphics.drawable.VectorDrawable
-import android.os.Build.VERSION.SDK_INT
+import android.os.Build.VERSION
 import android.os.Build.VERSION_CODES
 import android.util.Xml
 import androidx.annotation.DrawableRes
@@ -48,6 +48,8 @@ import androidx.core.graphics.component4
 import androidx.core.graphics.drawable.RoundedBitmapDrawable
 import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
 import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat
+import com.github.panpf.sketch.AndroidBitmap
+import com.github.panpf.sketch.ColorType
 import com.github.panpf.sketch.drawable.SketchDrawable
 import org.xmlpull.v1.XmlPullParser
 import org.xmlpull.v1.XmlPullParserException
@@ -85,7 +87,7 @@ internal fun Resources.getDrawableCompat(
 @SuppressLint("ResourceType")
 internal fun Context.getXmlDrawableCompat(resources: Resources, @XmlRes resId: Int): Drawable {
     // Modified from androidx.appcompat.widget.ResourceManagerInternal.
-    return if (SDK_INT >= 24) {
+    return if (VERSION.SDK_INT >= 24) {
         resources.getDrawableCompat(resId, theme)
     } else {
         // Find the XML's start tag.
@@ -112,22 +114,33 @@ internal fun Context.getXmlDrawableCompat(resources: Resources, @XmlRes resId: I
 /**
  * Drawable into new Bitmap. Each time a new bitmap is drawn
  *
- * @see com.github.panpf.sketch.core.android.test.util.DrawablesTest.testToNewBitmap
+ * @see com.github.panpf.sketch.core.android.test.util.DrawablesTest.testToBitmap
  */
-internal fun Drawable.toNewBitmap(
-    preferredConfig: Config? = null,
-    targetSize: Size? = null
+internal fun Drawable.toBitmap(
+    colorType: ColorType? = null,
+    colorSpace: ColorSpace? = null,
+    targetSize: Size? = null,
 ): Bitmap {
     val (oldLeft, oldTop, oldRight, oldBottom) = bounds
     val targetWidth = targetSize?.width ?: intrinsicWidth
     val targetHeight = targetSize?.height ?: intrinsicHeight
     setBounds(0, 0, targetWidth, targetHeight)
 
-    val bitmap: Bitmap = Bitmap.createBitmap(
-        /* width = */ targetWidth,
-        /* height = */ targetHeight,
-        /* config = */ preferredConfig.safeToSoftware(),
-    )
+    val bitmap: Bitmap = if (VERSION.SDK_INT >= VERSION_CODES.O && colorSpace != null) {
+        AndroidBitmap(
+            width = targetWidth,
+            height = targetHeight,
+            config = colorType.safeToSoftware(),
+            hasAlpha = true,
+            colorSpace = colorSpace
+        )
+    } else {
+        AndroidBitmap(
+            width = targetWidth,
+            height = targetHeight,
+            config = colorType.safeToSoftware()
+        )
+    }
     val canvas = Canvas(bitmap)
     draw(canvas)
 
@@ -160,16 +173,16 @@ fun Drawable.toLogString(): String = when {
     this is SketchDrawable -> toString()
     this is BitmapDrawable -> "BitmapDrawable(${bitmap.toLogString()})"
     this is RoundedBitmapDrawable -> "RoundedBitmapDrawable(drawable=${bitmap?.toLogString()})"
-    SDK_INT >= VERSION_CODES.P && this is AnimatedImageDrawable -> "AnimatedImageDrawable(${toSizeString()})"
+    VERSION.SDK_INT >= VERSION_CODES.P && this is AnimatedImageDrawable -> "AnimatedImageDrawable(${toSizeString()})"
     this is AnimatedVectorDrawableCompat -> "AnimatedVectorDrawableCompat(${toSizeString()})"
-    SDK_INT >= VERSION_CODES.LOLLIPOP && this is AnimatedVectorDrawable -> "AnimatedVectorDrawable(${toSizeString()})"
+    VERSION.SDK_INT >= VERSION_CODES.LOLLIPOP && this is AnimatedVectorDrawable -> "AnimatedVectorDrawable(${toSizeString()})"
     this is TransitionDrawable -> "TransitionDrawable(${toSizeString()})"
     this is ColorDrawable -> "ColorDrawable(${color})"
-    SDK_INT >= VERSION_CODES.Q && this is ColorStateListDrawable -> "ColorStateListDrawable(${toSizeString()})"
+    VERSION.SDK_INT >= VERSION_CODES.Q && this is ColorStateListDrawable -> "ColorStateListDrawable(${toSizeString()})"
     this is VectorDrawableCompat -> "VectorDrawableCompat(${toSizeString()})"
-    SDK_INT >= VERSION_CODES.LOLLIPOP && this is VectorDrawable -> "VectorDrawable(${toSizeString()})"
+    VERSION.SDK_INT >= VERSION_CODES.LOLLIPOP && this is VectorDrawable -> "VectorDrawable(${toSizeString()})"
     this is GradientDrawable -> "GradientDrawable(${toSizeString()})"
-    SDK_INT >= VERSION_CODES.M && this is DrawableWrapper -> "DrawableWrapper(drawable=${drawable?.toLogString()})"
+    VERSION.SDK_INT >= VERSION_CODES.M && this is DrawableWrapper -> "DrawableWrapper(drawable=${drawable?.toLogString()})"
     this is DrawableWrapperCompat -> "DrawableWrapperCompat(drawable=${drawable?.toLogString()})"
     else -> toString()
 }